% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_estimators.R
\name{ARD_SBM_regression}
\alias{ARD_SBM_regression}
\title{Regression Estimationg for Estimating the SBM given a set of cluster memberships}
\usage{
ARD_SBM_regression(traits, X, Z, constrained_opt = T, solver = "ECOS")
}
\arguments{
\item{traits}{Traits vector or matrix}

\item{X}{ARD responses}

\item{Z}{Cluster memberships}

\item{constrained_opt}{Use constrained optimization}

\item{solver}{Solver to use in CVX if using constrained optimization}
}
\value{
Estimated SBM P matrix
}
\description{
Regression Estimationg for Estimating the SBM given a set of cluster memberships
}
