% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_treatments.R
\name{binomialTreatment}
\alias{binomialTreatment}
\title{Sample A Binomial Treatment}
\usage{
binomialTreatment(n, p)
}
\arguments{
\item{n}{Number of samples}

\item{p}{Probability of Treatment Assignment}
}
\value{
Treatment Assignments
}
\description{
Sample A Binomial Treatment
}
