% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_leiden_K}
\alias{cluster_leiden_K}
\title{Use Leiden Clustering to compute the clusters}
\usage{
cluster_leiden_K(
  G,
  K = 4,
  adj_mode = "undirected",
  max_resolution_parameter = 10,
  n_iterations = 5
)
}
\arguments{
\item{G}{Adjacency matrix}

\item{K}{Number of clusters}

\item{adj_mode}{Only 'undirected' is supported}

\item{max_resolution_parameter}{Maximum resolution parameter (tuning parameter for finding clusters)}

\item{n_iterations}{Number of iterations for Leiden algorithm}
}
\value{
Cluster membership
}
\description{
Use Leiden Clustering to compute the clusters
}
