% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clusterARDHclust}
\alias{clusterARDHclust}
\title{Computes Hierarchical Clustering of the ARD responses}
\usage{
clusterARDHclust(X, K, alg = "complete")
}
\arguments{
\item{X}{ARD responses, should be normalized}

\item{K}{Number of clusters}

\item{alg}{Algorithm used For merging}
}
\value{
Hierarchical Clustering
}
\description{
Computes Hierarchical Clustering of the ARD responses
}
