% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clusterARDKmeans}
\alias{clusterARDKmeans}
\title{Computes K-means clustering for ARD responses.}
\usage{
clusterARDKmeans(X, K, alg = "Lloyd")
}
\arguments{
\item{X}{ARD responses, should be normalized}

\item{K}{Number of clusters}

\item{alg}{Algorithm Used}
}
\value{
K-means clustering
}
\description{
Computes K-means clustering for ARD responses.
}
