% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_treatments.R
\name{clusterTreatment}
\alias{clusterTreatment}
\title{Cluster Treatment Allocation}
\usage{
clusterTreatment(Z, p.treat = 1/2)
}
\arguments{
\item{Z}{Cluster Membership}

\item{p.treat}{Probability of Fully Treated Clusters}
}
\value{
Treatment Assignments
}
\description{
Cluster Treatment Allocation
}
