% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeARD}
\alias{computeARD}
\title{Compute Synthetic ARD from a list of discrete traits}
\usage{
computeARD(traits, G)
}
\arguments{
\item{traits}{Vector or matrix of traits}

\item{G}{Adjacency Matrix}
}
\value{
ARD responses
}
\description{
Compute Synthetic ARD from a list of discrete traits
}
