% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeding.R
\name{ContagionOptimalSeeding}
\alias{ContagionOptimalSeeding}
\title{Optimal Seeding for Contagion}
\usage{
ContagionOptimalSeeding(
  SBM,
  n_seeds = 2,
  num_sim = 2000,
  contagion_type = "simple",
  complex_mean = NULL,
  complex_sd = NULL,
  ...
)
}
\arguments{
\item{SBM}{Model Parameters for the SBM}

\item{n_seeds}{Number of total seeds}

\item{num_sim}{Number of Simulations to conduct for each seed combination}

\item{contagion_type}{Simple or Complex}

\item{complex_mean}{Complex Contagion Thresholds mean}

\item{complex_sd}{Complex Contagion Thresholds sd}

\item{...}{Additional arguments}
}
\value{
Dataframe with results of the optimal seeding
}
\description{
Optimal Seeding for Contagion
}
