% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_features.R
\name{degree_ratio}
\alias{degree_ratio}
\title{Degree Ratio}
\usage{
degree_ratio(data, G)
}
\arguments{
\item{data}{data frame}

\item{G}{Adjacency matrix}
}
\value{
degree_ratio
}
\description{
Degree Ratio
}
