% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_features.R
\name{frac_treated_neighbors}
\alias{frac_treated_neighbors}
\title{Fraction of Treated Neighbors}
\usage{
frac_treated_neighbors(data, G, treatment_column = "A")
}
\arguments{
\item{data}{data frame}

\item{G}{Adjacency matrix}

\item{treatment_column}{# which column name indicates the treatment}
}
\value{
fraction of treated neighbors
}
\description{
Fraction of Treated Neighbors
}
