% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Generate_G_set}
\alias{Generate_G_set}
\title{Generate Set Of SBM Draws}
\usage{
Generate_G_set(SBM, L = 1000)
}
\arguments{
\item{SBM}{Stochastic Block Model}

\item{L}{Number of draws}
}
\value{
List of Adjacency Matrices
}
\description{
Generate Set Of SBM Draws
}
