% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateSBM}
\alias{generateSBM}
\title{Simulate Draw of an SBM}
\usage{
generateSBM(n, P, PI, Z)
}
\arguments{
\item{n}{number of nodes to simulate}

\item{P}{Probability matrix}

\item{PI}{Proportions of each group}

\item{Z}{Group membership}
}
\value{
Adjacency matrix and group membership
}
\description{
Simulate Draw of an SBM
}
