% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelSwitching}
\alias{labelSwitching}
\title{Label swapping to align an estimated labels with a reference set}
\usage{
labelSwitching(Z.ref, Z.hat)
}
\arguments{
\item{Z.ref}{Reference labels}

\item{Z.hat}{Estimated labels}
}
\value{
Aligned Estimated labels
}
\description{
Label swapping to align an estimated labels with a reference set
}
