% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelSwitchingPermutations}
\alias{labelSwitchingPermutations}
\title{Label swapping to align an estimated labels with a reference set}
\usage{
labelSwitchingPermutations(Z.ref, Z.hat)
}
\arguments{
\item{Z.ref}{Reference labels}

\item{Z.hat}{Estimated labels}
}
\value{
permutations
}
\description{
Label swapping to align an estimated labels with a reference set
}
