% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_design.R
\name{linear_variance_function}
\alias{linear_variance_function}
\title{Variance Function Generator}
\usage{
linear_variance_function(
  A,
  phi,
  formula,
  data,
  network_feature_list,
  G_set,
  Sigma,
  verbose = T
)
}
\arguments{
\item{A}{Treatment Assignment}

\item{phi}{Coefficients for the linear combination}

\item{formula}{Formula for the regression model}

\item{data}{Observed data frame}

\item{network_feature_list}{List of network features to include in the model}

\item{G_set}{Set of samples of the graph}

\item{Sigma}{Correlation matrix of the additive noise}

\item{verbose}{Print progress}
}
\value{
Variance of the linear combination of coefficients
}
\description{
Variance Function Generator
}
