% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_design.R
\name{linear_variance_function_factory}
\alias{linear_variance_function_factory}
\title{Function Factory for Bayesian Optimization of the linear variance function}
\usage{
linear_variance_function_factory(
  phi,
  formula,
  data,
  network_feature_list,
  G_set,
  clusters,
  Sigma
)
}
\arguments{
\item{phi}{Coefficients for the linear combination}

\item{formula}{Formula for the regression model}

\item{data}{Observed data frame}

\item{network_feature_list}{List of network features to include in the model}

\item{G_set}{Set of samples of the graph}

\item{Sigma}{Correlation matrix of the additive noise}
}
\value{
Variance where one can compute this using a function of group saturations tau
}
\description{
Function Factory for Bayesian Optimization of the linear variance function
}
