% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mclustBIC}
\alias{mclustBIC}
\title{Importing mclustBIC function}
\usage{
mclustBIC(
  data,
  G = NULL,
  modelNames = NULL,
  prior = NULL,
  control = emControl(),
  initialization = list(hcPairs = NULL, subset = NULL, noise = NULL),
  Vinv = NULL,
  warn = mclust.options("warn"),
  x = NULL,
  verbose = interactive(),
  ...
)
}
\description{
Importing mclustBIC function
}
