% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_features.R
\name{network_features}
\alias{network_features}
\title{Network Features}
\usage{
network_features(data, G, feat_list = list(degree = degree))
}
\arguments{
\item{data}{base data frame}

\item{G}{Adjacency matrix}

\item{feat_list}{List of network feature functions}
}
\value{
data frame with network features
}
\description{
Network Features
}
