% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_clusters_from_prob_matrix}
\alias{sample_clusters_from_prob_matrix}
\title{Sample From Probability Matrix}
\usage{
sample_clusters_from_prob_matrix(probabilities)
}
\arguments{
\item{probabilities}{}
}
\value{
Sampled clusters
}
\description{
Sample From Probability Matrix
}
