% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_treatments.R
\name{saturationRandomizationTreatment}
\alias{saturationRandomizationTreatment}
\title{Cluster Randomization}
\usage{
saturationRandomizationTreatment(Z, levels)
}
\arguments{
\item{Z}{Cluster Membership}

\item{levels}{Saturation levels for each of the clusters assigned in Z}
}
\value{
Treatment Assignments
}
\description{
Cluster Randomization
}
