% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_estimators.R
\name{SBM_bootstrap_ARD}
\alias{SBM_bootstrap_ARD}
\title{Estimate draws of the ARD model from trait responses.}
\usage{
SBM_bootstrap_ARD(
  traits,
  X,
  K,
  L = 1000,
  alg = "ward.D",
  frac_observed = 1,
  verbose = T,
  cluster_traits_quantile = 0.75
)
}
\arguments{
\item{traits}{Self traits}

\item{X}{ARD responses}

\item{K}{Number of clusters}

\item{L}{Number of Bootstrap iterations}

\item{alg}{Algorithm Used for Agglomerative Clustering}

\item{frac_observed}{Correction if there is a fraction of the observed data}

\item{verbose}{Print progress}

\item{cluster_traits_quantile}{Quantile for total trait threshold}
}
\value{
Bootstrap draws of the Stochastic Blockmodel
}
\description{
Estimate draws of the ARD model from trait responses.
}
