% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_estimators.R
\name{SBM_estimate_ARD}
\alias{SBM_estimate_ARD}
\title{Estimate ARD blocks from trait responses.}
\usage{
SBM_estimate_ARD(
  traits,
  X,
  K,
  alg = "ward.D",
  Z.ref = NULL,
  constrained_opt = T,
  solver = "ECOS",
  cluster_traits_quantile = 0.75
)
}
\arguments{
\item{traits}{Self traits}

\item{X}{ARD responses}

\item{K}{Number of clusters}

\item{alg}{Which algorithm will be used for agglomerative clustering}

\item{Z.ref}{Reference cluster labels}

\item{constrained_opt}{Use constrained optimization}

\item{solver}{Solver to use in CVX if using constrained optimization}

\item{cluster_traits_quantile}{Quantile for total trait threshold}
}
\value{
Stochastic block model (P and Z)
}
\description{
Estimate ARD blocks from trait responses.
}
