% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_estimators.R
\name{SBM_estimate_subgraph}
\alias{SBM_estimate_subgraph}
\title{Estimate SBM using subgraph and cluster labels}
\usage{
SBM_estimate_subgraph(Z, G)
}
\arguments{
\item{Z}{cluster memberships}

\item{G}{adjacency matrix}
}
\value{
Stochastic block model (P and Z)
}
\description{
Estimate SBM using subgraph and cluster labels
}
