% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{SBM_glm}
\alias{SBM_glm}
\title{GLM Fit from sample}
\usage{
SBM_glm(
  formula,
  data,
  family = quasibinomial(),
  SBM,
  network_feature_list = list(degree = degree, frac.treated = frac_treated_neighbors),
  B = 1000,
  EM_threshold = 1e-06,
  max_iter = 100
)
}
\arguments{
\item{formula}{Regression formula}

\item{data}{Dataset}

\item{family}{GLM family}

\item{SBM}{Model Parameters for the SBM}

\item{network_feature_list}{List of network features to include in the model}

\item{B}{Number of Feature Samples}

\item{EM_threshold}{Tuning threshold for EM algorithm}

\item{max_iter}{Maximum number of iterations for the EM algorithm}
}
\value{
Glm model for binary data
}
\description{
GLM Fit from sample
}
