% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{SBM_lm}
\alias{SBM_lm}
\title{Generates Draws from the SBM and estimates a linear model using the averaged coefficients}
\usage{
SBM_lm(
  formula,
  data,
  SBM,
  network_feature_list = list(degree = degree, frac.treated = frac_treated_neighbors),
  B = 1000,
  verbose = T
)
}
\arguments{
\item{formula}{Regression formula}

\item{data}{Dataset}

\item{SBM}{Model Parameters for the SBM}

\item{network_feature_list}{List of network features to include in the model}

\item{B}{Number of simulations}

\item{verbose}{Print progress}
}
\value{
List with the model and the mean data
}
\description{
Generates Draws from the SBM and estimates a linear model using the averaged coefficients
}
