% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_estimators.R
\name{SBM_uncertainty_ARD}
\alias{SBM_uncertainty_ARD}
\title{Estimate draws of the ARD model from trait responses.}
\usage{
SBM_uncertainty_ARD(
  traits,
  X,
  K,
  L = 1000,
  alg = "ward.D",
  cluster_traits_quantile = 0.75
)
}
\arguments{
\item{traits}{Self traits}

\item{X}{ARD responses}

\item{K}{Number of clusters}

\item{L}{Number of Bootstrap iterations}

\item{alg}{Algorithm Used for Agglomerative Clustering}

\item{cluster_traits_quantile}{Quantile for total trait threshold}
}
\value{
List of Draws from the Stochastic Blockmodel
}
\description{
Estimate draws of the ARD model from trait responses.
}
