% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMNetReg-package.R
\docType{package}
\name{SBMNetReg-package}
\alias{SBMNetReg}
\alias{SBMNetReg-package}
\title{SBMNetReg: Compute Regression Estimates Using Stochastic Blockmodels For Partial Network Data}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
