% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{simLinearModel}
\alias{simLinearModel}
\title{Simulate A Linear Model From The Features}
\usage{
simLinearModel(data, fmla, coef, sd = 1, noise_covar = NULL)
}
\arguments{
\item{data}{Input Data}

\item{fmla}{Formula for the model}

\item{coef}{Coefficients}

\item{sd}{Standard Deviation for iid data}

\item{noise_covar}{Covariance matrix for the noise for non-iid}
}
\value{
List with the outcomes, the features and the noise covariance
}
\description{
Simulate A Linear Model From The Features
}
