% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{simLogisticModel}
\alias{simLogisticModel}
\title{Simulate a simple logistic model}
\usage{
simLogisticModel(data, fmla, coef)
}
\arguments{
\item{data}{Input Data}

\item{fmla}{Formula for the model}

\item{coef}{Coefficients}
}
\value{
List with the outcomes, the features and the noise covariance
}
\description{
Simulate a simple logistic model
}
