% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_features.R
\name{treated_neighbors}
\alias{treated_neighbors}
\title{Number of Treated Neighbors}
\usage{
treated_neighbors(data, G, treatment_column = "A")
}
\arguments{
\item{data}{data frame}

\item{G}{Adjacency matrix}

\item{treatment_column}{# which column name indicates the treatment}
}
\value{
Number of treated neighbors
}
\description{
Number of Treated Neighbors
}
